src/security-service/requirements.txt
# =============================================================================
# Security Service - Dependencies
# =============================================================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
supabase>=2.11.0

# Data Validation
pydantic==2.10.0
pydantic-settings==2.11.0

# HTTP Client (for Supabase internal calls)
httpx>=0.28.1,<1.0.0

# JWT & Cryptography
PyJWT>=2.8.0,<3.0.0

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6

# Development & Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
ruff==0.1.9
